Refactored fixtures file
