+++ apps/ledger-live-desktop/tests/specs/speculos/add.account.spec.ts
@@ -1,7 +1,7 @@
 import test from "../../fixtures/common";
 import { specs } from "../../utils/speculos";
 import { Currency } from "../../enum/Currency";
-import { addTmsLink } from "tests/fixtures/common";
+import { addTmsLink } from "tests/utils/allureUtils";
 
 const currencies: Currency[] = [
   Currency.BTC,
