+++ apps/ledger-live-desktop/tests/specs/speculos/receive.address.spec.ts
@@ -1,7 +1,7 @@
 import test from "../../fixtures/common";
 import { Account } from "../../enum/Account";
 import { specs } from "../../utils/speculos";
-import { addTmsLink } from "tests/fixtures/common";
+import { addTmsLink } from "tests/utils/allureUtils";
 
 const accounts: Account[] = [
   // Derivation path is updated when account receive money
